requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aioice==0.9.0
aiortc==1.9.0
aiosignal==1.3.1
altair==5.5.0
annotated-types==0.7.0
anyio==4.7.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.4
async-timeout==4.0.3
asyncer==0.0.7
attrs==24.2.0
autocommand==2.2.2
av==12.3.0
babel==2.16.0
backports.tarfile==1.2.0
beautifulsoup4==4.12.3
bidict==0.23.1
bleach==6.2.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
chainlit==1.3.2
charset-normalizer==3.4.0
chevron==0.14.0
click==8.1.8
colorama==0.4.6
comm==0.2.2
comtypes==1.4.8
cryptography==44.0.0
dataclasses-json==0.6.7
debugpy==1.8.9
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.15
distro==1.9.0
dnspython==2.7.0
exceptiongroup==1.2.2
executing==2.1.0
fastapi==0.115.6
fastjsonschema==2.21.1
filelock==3.16.1
filetype==1.2.0
fpdf==1.7.2
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2024.12.0
gitdb==4.0.11
GitPython==3.1.43
google-crc32c==1.6.0
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpcio==1.68.1
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
httpx-sse==0.4.0
idna==3.10
ifaddr==0.2.0
importlib_metadata==8.5.0
importlib_resources==6.4.0
inflect==7.3.1
ipykernel==6.29.5
ipython==8.30.0
isoduration==20.11.0
itsdangerous==2.2.0
jaraco.collections==5.1.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jaraco.text==4.0.0
jedi==0.19.2
Jinja2==3.1.5
jiter==0.8.2
json5==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.11.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
langchain==0.3.11
langchain-community==0.3.11
langchain-core==0.3.24
langchain-openai==0.2.12
langchain-text-splitters==0.3.2
langgraph==0.2.59
langgraph-checkpoint==2.0.9
langgraph-sdk==0.1.48
langsmith==0.2.3
Lazify==0.4.0
literalai==0.0.623
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
more-itertools==10.5.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
narwhals==1.19.1
nbclient==0.10.2
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
notebook==7.3.2
notebook_shim==0.2.4
numpy==1.26.4
openai==1.57.3
opentelemetry-api==1.28.2
opentelemetry-exporter-otlp==1.28.2
opentelemetry-exporter-otlp-proto-common==1.28.2
opentelemetry-exporter-otlp-proto-grpc==1.28.2
opentelemetry-exporter-otlp-proto-http==1.28.2
opentelemetry-instrumentation==0.49b2
opentelemetry-proto==1.28.2
opentelemetry-sdk==1.28.2
opentelemetry-semantic-conventions==0.49b2
orjson==3.10.12
overrides==7.7.0
packaging==23.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pillow==11.0.0
pip-chill==1.0.3
platformdirs==4.2.2
prometheus_client==0.21.1
prompt_toolkit==3.0.48
propcache==0.2.1
protobuf==5.29.1
psutil==6.1.0
pure_eval==0.2.3
pyarrow==18.1.0
PyAudio==0.2.14
pycparser==2.22
pydantic==2.10.1
pydantic-settings==2.6.1
pydantic_core==2.27.1
pydeck==0.9.1
pydub==0.25.1
pyee==12.1.1
Pygments==2.18.0
PyJWT==2.10.1
pylibsrtp==0.10.0
pyOpenSSL==24.3.0
pypiwin32==223
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.10.1
python-json-logger==3.2.1
python-multipart==0.0.9
python-socketio==5.11.4
pyttsx3==2.98
pytz==2024.2
pywin32==308
pywinpty==2.0.14
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.22.3
Send2Trash==1.8.3
simple-websocket==1.1.0
six==1.17.0
smmap==5.0.1
sniffio==1.3.1
soundfile==0.12.1
soupsieve==2.6
SpeechRecognition==3.12.0
SQLAlchemy==2.0.36
stack-data==0.6.3
starlette==0.41.3
streamlit==1.41.1
streamlit-webrtc==0.47.9
streamlit_mic_recorder==0.0.8
sympy==1.13.1
syncer==2.0.3
tenacity==9.0.0
terminado==0.18.1
tiktoken==0.8.0
tinycss2==1.4.0
toml==0.10.2
tomli==2.2.1
torch==2.5.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typeguard==4.4.1
types-python-dateutil==2.9.0.20241206
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
uptrace==1.28.2
uri-template==1.3.0
urllib3==2.2.3
uvicorn==0.25.0
watchdog==6.0.0
watchfiles==0.20.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
whisper==1.1.10
wrapt==1.17.0
wsproto==1.2.0
yarl==1.18.3
zipp==3.21.0

-e .